remove duplicate code in the schema-registry module
fixes #6
