+++ embedded-kafka/src/main/scala/net/manub/embeddedkafka/ops/embeddedKafkaOps.scala
@@ -1,0 +1,34 @@
+package net.manub.embeddedkafka.ops
+
+import net.manub.embeddedkafka.{EmbeddedKafkaConfig, EmbeddedServer}
+
+/**
+  * Combined trait with all Ops traits.
+  *
+  * @tparam C an [[EmbeddedKafkaConfig]]
+  * @tparam S an [[EmbeddedServer]]
+  */
+private[embeddedkafka] trait EmbeddedKafkaOps[
+    C <: EmbeddedKafkaConfig, S <: EmbeddedServer]
+    extends AdminOps[C]
+    with ConsumerOps[C]
+    with ProducerOps[C]
+    with ZooKeeperOps
+    with KafkaOps
+
+/**
+  * [[EmbeddedKafkaOps]] extension relying on [[RunningServersOps]]
+  * for keeping track of running [[EmbeddedServer]]s.
+  *
+  * @tparam C an [[EmbeddedKafkaConfig]]
+  * @tparam S an [[EmbeddedServer]]
+  *
+  * @see [[RunningServersOps]]
+  */
+private[embeddedkafka] trait RunningEmbeddedKafkaOps[
+    C <: EmbeddedKafkaConfig, S <: EmbeddedServer]
+    extends EmbeddedKafkaOps[C, S]
+    with RunningServersOps
+    with ServerStarter[C, S]
+    with RunningZooKeeperOps
+    with RunningKafkaOps
