+++ embedded-kafka/src/test/scala/net/manub/embeddedkafka/embeddedKafkaSpecSupport.scala
@@ -44,13 +44,6 @@ abstract class EmbeddedKafkaSpecSupport
     expectMsg(1 second, Connection.Success)
   }
 
-  def schemaRegistryIsAvailable(schemaRegistryPort: Int = 6002): Unit = {
-    system.actorOf(
-      TcpClient.props(new InetSocketAddress("localhost", schemaRegistryPort),
-                      testActor))
-    expectMsg(1 second, Connection.Success)
-  }
-
   def zookeeperIsAvailable(zookeeperPort: Int = 6000): Unit = {
     system.actorOf(
       TcpClient.props(new InetSocketAddress("localhost", zookeeperPort),
@@ -61,13 +54,6 @@ abstract class EmbeddedKafkaSpecSupport
   def kafkaIsNotAvailable(kafkaPort: Int = 6001): Unit = {
     system.actorOf(
       TcpClient.props(new InetSocketAddress("localhost", kafkaPort), testActor))
-    expectMsg(1 second, Connection.Failure)
-  }
-
-  def schemaRegistryIsNotAvailable(schemaRegistryPort: Int = 6002): Unit = {
-    system.actorOf(
-      TcpClient.props(new InetSocketAddress("localhost", schemaRegistryPort),
-                      testActor))
     expectMsg(1 second, Connection.Failure)
   }
