+++ kafka-streams/src/main/scala/net/manub/embeddedkafka/streams/EmbeddedStreamsConfig.scala
@@ -1,0 +1,43 @@
+package net.manub.embeddedkafka.streams
+
+import java.nio.file.Files
+
+import net.manub.embeddedkafka.EmbeddedKafkaConfig
+import org.apache.kafka.clients.consumer.{ConsumerConfig, OffsetResetStrategy}
+import org.apache.kafka.streams.StreamsConfig
+
+/**
+  * Kafka Streams config generator.
+  *
+  * @tparam C an [[EmbeddedKafkaConfig]]
+  */
+private[embeddedkafka] trait EmbeddedStreamsConfig[C <: EmbeddedKafkaConfig] {
+
+  protected[embeddedkafka] def baseStreamConfig(streamName: String)(
+      implicit kafkaConfig: C): Map[String, AnyRef] = Map(
+    StreamsConfig.APPLICATION_ID_CONFIG -> streamName,
+    StreamsConfig.BOOTSTRAP_SERVERS_CONFIG -> s"localhost:${kafkaConfig.kafkaPort}",
+    StreamsConfig.STATE_DIR_CONFIG -> Files
+      .createTempDirectory(streamName)
+      .toString,
+    // force stream consumers to start reading from the beginning so as not to lose messages
+    ConsumerConfig.AUTO_OFFSET_RESET_CONFIG -> OffsetResetStrategy.EARLIEST.toString.toLowerCase
+  )
+
+  /**
+    * Returns a map of config parameters for running Kafka Streams.
+    *
+    * @param streamName  the name of the stream. It will be used as the Application ID
+    * @param extraConfig any additional configuration. If the keys are already defined
+    *                    in the default they will be overwritten with this
+    * @param kafkaConfig the Kafka test configuration
+    * @return
+    */
+  def config(streamName: String, extraConfig: Map[String, AnyRef])(
+      implicit kafkaConfig: C): Map[String, AnyRef] =
+    baseStreamConfig(streamName) ++ extraConfig
+
+}
+
+final class EmbeddedStreamsConfigImpl
+    extends EmbeddedStreamsConfig[EmbeddedKafkaConfig]
