+++ schema-registry/src/main/scala/net.manub.embeddedkafka/schemaregistry/streams/EmbeddedKafkaStreamsWithSchemaRegistryAllInOne.scala
@@ -1,37 +1,16 @@
 package net.manub.embeddedkafka.schemaregistry.streams
 
-import net.manub.embeddedkafka.Consumers
 import net.manub.embeddedkafka.schemaregistry.EmbeddedKafkaConfigWithSchemaRegistry
-import org.apache.kafka.clients.consumer.KafkaConsumer
-import org.apache.kafka.streams.Topology
+import net.manub.embeddedkafka.streams.EmbeddedKafkaStreamsAllInOneSupport
 
-// TODO: need to find a better way of not duplicating this code from the kafka-streams module
-
-/** Convenience trait for testing Kafka Streams with ScalaTest.
-  * It exposes `EmbeddedKafkaStreams.runStreams` as well as `Consumers` api
-  * for easily creating and querying consumers in tests.
+/** Convenience trait exposing [[EmbeddedKafkaStreamsWithSchemaRegistry.runStreams]]
+  * as well as [[net.manub.embeddedkafka.Consumers]] api for easily creating and
+  * querying consumers.
   *
-  * @see [[Consumers]]
   * @see [[EmbeddedKafkaStreamsWithSchemaRegistry]]
+  * @see [[net.manub.embeddedkafka.Consumers]]
   */
 trait EmbeddedKafkaStreamsWithSchemaRegistryAllInOne
-    extends EmbeddedKafkaStreamsWithSchemaRegistry
-    with Consumers {
-
-  /** Run Kafka Streams while offering a String-based consumer for easy testing of stream output.
-    *
-    * @param topicsToCreate the topics that should be created. Usually these should be the topics
-    *                       that the Streams-under-test use for inputs and outputs. They need to be
-    *                       created before running the streams and
-    *                       this is automatically taken care of.
-    * @param topology       the streams topology that will be instantiated
-    * @param block          the block of testing code that will be executed by passing the simple
-    *                       String-based consumer.
-    * @return the result of the testing code
-    */
-  def runStreamsWithStringConsumer(
-      topicsToCreate: Seq[String],
-      topology: Topology)(block: KafkaConsumer[String, String] => Any)(
-      implicit config: EmbeddedKafkaConfigWithSchemaRegistry): Any =
-    runStreams(topicsToCreate, topology)(withStringConsumer[Any](block))(config)
-}
+    extends EmbeddedKafkaStreamsAllInOneSupport[
+      EmbeddedKafkaConfigWithSchemaRegistry]
+    with EmbeddedKafkaStreamsWithSchemaRegistry
